Move function to its own package
